+++ changelog.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated the helpers to be installed from the CDN
 
+### Fixed
+
+- Build script was not updating the rigth `@build.version` so it was never updated in the bvm and installer.'
+
 ## [1.14.0] - 2025-06-24
 
 ### Added